% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{extractSSB}
\alias{extractSSB}
\alias{extractRec}
\alias{extractFbar}
\alias{extractZatage}
\alias{extractDevs}
\title{Extracts derived quantities from SS output}
\usage{
extractSSB(out)

extractRec(out)

extractFbar(out)

extractZatage(out)

extractDevs(out)
}
\arguments{
\item{out}{A list as returned by r4ss::SS_output.}
}
\value{
An \emph{FLQuant} object of the requested quantity.
}
\description{
A number of derived quantities are available in the \emph{Report.sso} file, and
are useful for checking that the generated FLR objects lead to the same
values. These functions are called on the list returned by \emph{readOutputss3}
or r4ss::SS_output, and extract the yearly values from the following rows
in the \emph{derived_quants} data.frame:
}
\details{
\itemize{
\item \emph{extractSSB}: \code{SSB_y} for y between \code{startyr} and \code{endyr}.
\item \emph{extractRec}: \code{Recr_y} for y between \code{startyr} and \code{endyr}.
\item \emph{extractFbar}: \code{F_y} for y between \code{startyr} and \code{endyr}.
\item \emph{extractZatage}: \code{Z_ay} for y between \code{startyr} and \code{endyr} and ages but last.
}

For 2 sex models (\code{nsexes}), \emph{extractRec} will return a two-unit \emph{FLQuant},
with the \code{Recr_y} values split according to the recruitment sex ratio. This
is extracted from \code{recruitment_dist[[1]][, "Frac/sex"]}.

The value returned by \code{extractFbar} is the actual mean F over the age range,
and the value in \code{derived_quants} is corrected according to \code{F_report_basis}
for models where F reporting basis is F/FMSY.
}
\examples{
out <- readOutputss3(system.file("ext-data", "herring", package="ss3om"))

extractSSB(out)

extractFbar(out)
}
