% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{nameGrid}
\alias{nameGrid}
\title{nameGrid}
\usage{
nameGrid(df, dir, from = 1)
}
\arguments{
\item{df}{Model grid data.frame}

\item{dir}{Folder name}

\item{from}{Starting number}
}
\description{
Creates folder names from a 'grid' df of scenarios
}
