% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLSRss3}
\alias{readFLSRss3}
\title{A function to read the stock-recruit relationships from an SS3 run into an \code{FLSR} object}
\usage{
readFLSRss3(dir, birthseas = out$birthseas, ...)
}
\arguments{
\item{dir}{Directory containing the SS3 output files}

\item{birthseas}{The birthseasons for this stock as a numeric vector.}

\item{...}{Any other argument to be passed to \code{r4ss::SS_output}}
}
\value{
An object of class \link[FLCore:FLStock]{FLStock}
}
\description{
A function to read the stock-recruit relationships from an SS3 run into an \code{FLSR} object
}
\references{
Methot RD Jr, Wetzel CR (2013) Stock Synthesis: A biological and statistical
framework for fish stock assessment and fishery management.
Fisheries Research 142: 86-99.
}
\seealso{
\link{FLComp}
}
\author{
Iago Mosqueira, EC JRC
}
\keyword{classes}
