% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSS3.R
\name{readFLSss3}
\alias{readFLSss3}
\title{A function to read SS3 results as an FLStock object}
\usage{
readFLSss3(dir, birthseas = out$spawnseas, name = "",
  desc = paste(out$inputs$repfile, out$SS_version, sep = " - "))
}
\arguments{
\item{dir}{Directory holding the SS3 output files}

\item{birthseas}{Birth seasons for this stock, defaults to spawnseas}

\item{name}{Name of the output object to fil the name slot}

\item{desc}{Description of the output object to fill the desc slot}
}
\value{
An object of class \code{\link{FLStock}}
}
\details{
Aliquam sagittis feugiat felis eget consequat. Praesent eleifend dolor massa, 
vitae faucibus justo lacinia a. Cras sed erat et magna pharetra bibendum quis in 
mi. Sed sodales mollis arcu, sit amet venenatis lorem fringilla vel. Vivamus vitae 
ipsum sem. Donec malesuada purus at libero bibendum accumsan. Donec ipsum sapien, 
feugiat blandit arcu in, dapibus dictum felis.

Methot RD Jr, Wetzel CR (2013) Stock Synthesis: A biological and statistical
framework for fish stock assessment and fishery management.
Fisheries Research 142: 86–99.
}
\examples{

}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
