% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3slots.R, R/ss3slots30.R
\name{ss3slot}
\alias{ss3slot}
\alias{ss3index}
\alias{ss3index.res}
\alias{ss3index.var}
\alias{ss3index.q}
\alias{ss3sel.pattern}
\alias{ss3wt}
\alias{ss3mat}
\alias{ss3m}
\alias{ss3n}
\alias{ss3catch}
\alias{ss3mat30}
\alias{ss3m30}
\alias{ss3n30}
\alias{ss3catch30}
\title{Functions to convert SS3 output into FLQuant(s)}
\usage{
ss3index(cpue, fleets)

ss3index.res(cpue, fleets)

ss3index.var(cpue, fleets)

ss3index.q(cpue, fleets)

ss3sel.pattern(selex, years, fleets, morphs, factor = "Asel2")

ss3wt(endgrowth, dmns, birthseas)

ss3mat(endgrowth, dmns, birthseas, option = 3)

ss3m(endgrowth, dmns, birthseas)

ss3n(n, dmns, birthseas)

ss3catch(catage, wtatage, dmns, birthseas, idx)

ss3mat30(endgrowth, dmns, spawnseas, option = 3)

ss3m30(endgrowth, dmns, birthseas)

ss3n30(n, dmns)

ss3catch30(catage, wtatage, dmns, birthseas, idx)
}
\arguments{
\item{cpue}{A data frame obtained from SS_output$cpue.}

\item{fleets}{Named vector of fleets (numeric) codes}

\item{selex}{A data frame obtained from SS_output$ageselex.}

\item{years}{Vector of years for which the index applies}

\item{morphs}{Vector of morphs to use}

\item{endgrowth}{A data frame obtained from SS_output$endgrowth.}

\item{dmns}{dimnames of the output object, usually obatined using \code{getDimnames}.}

\item{birthseas}{The birthseasons for this stock as a numeric vector.}

\item{n}{A data frame obtained from SS_output$natage.}

\item{catage}{A data frame obtained from SS_output$catage.}

\item{wtatage}{A data frame obtained from SS_output$endgrowth but subset for \code{birthseas} and \verb{RetWt:_idx}.}

\item{idx}{The fishing fleets, as in \code{SS_output$fleet_ID[SS_output$IsFishFleet]}.}
}
\value{
An FLQuant or FLQuants object, depending on the converted data structure
}
\description{
A series of auxiliary functions that convert one or more elements, typically
of class \code{data.frame}. in the list returned by \code{r4ss::SS_output} into particular
FLQuant or FLQuants objects.
}
\details{
\itemize{
\item \code{ss3index} returns the \code{index} slot of each survey/CPUE fleet.
}

\itemize{
\item \code{ss3index.res} returns the \code{index.res} slot of each survey/CPUE fleet.
}

\itemize{
\item \code{ss3index.var} returns the \code{index.var} slot of each survey/CPUE fleet.
}

\itemize{
\item \code{ss3index.q} returns the \code{index.q} slot of each survey/CPUE fleet.
}

\itemize{
\item \code{ss3sel.pattern} returns the \code{sel.pattern} slot of each survey/CPUE fleet.
}

\itemize{
\item \code{ss3wt} returns the \code{stock.wt} slot.
}

\itemize{
\item \code{ss3mat} returns the \code{mat} slot.
}

\itemize{
\item \code{ss3m} returns the \code{m} slot.
}

\itemize{
\item \code{ss3m} returns the \code{m} slot.
}

\itemize{
\item \code{ss3catch} currently returns the \code{landings.n} slot, equal to \code{catch.n} as discards are not being parsed.
}

\itemize{
\item \code{ss3mat30} returns the \code{mat} slot.
}

\itemize{
\item \code{ss3m} returns the \code{m} slot.
}

\itemize{
\item \code{ss3n30} returns the \code{stock.n} slot.
}

\itemize{
\item \code{ss3catch} currently returns the \code{landings.n} slot, equal to \code{catch.n} as discards are not being parsed.
}
}
\seealso{
\code{\link{FLQuant}} \code{\link{readFLSss3}}
}
\author{
Iago Mosqueira, EC JRC D02
}
\keyword{classes}
