% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSS3.R
\name{readFLIBss3}
\alias{readFLIBss3}
\title{A function to read the CPUE series from an SS3 run into anm FLIndex object}
\usage{
readFLIBss3(dir, fleets)
}
\arguments{
\item{dir}{Directory holding the SS3 output files}

\item{name}{Name of the output object to fil the name slot}

\item{desc}{Description of the output object to fill the desc slot}
}
\value{
An object of class \code{\link{FLStock}}
}
\description{
Methot RD Jr, Wetzel CR (2013) Stock Synthesis: A biological and statistical
framework for fish stock assessment and fishery management.
Fisheries Research 142: 86–99.
}
\examples{
 
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
