% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readFLSss3}
\alias{readFLSss3}
\title{A function to read SS3 results as an FLStock object}
\usage{
readFLSss3(
  dir,
  repfile = "Report.sso",
  compfile = "CompReport.sso",
  wtatage = out$wtatage_switch,
  ...
)
}
\arguments{
\item{dir}{Directory holding the SS3 output files}

\item{...}{Any other argument to be passed to \code{r4ss::SS_output}}

\item{birthseas}{Birth seasons for this stock, defaults to spawnseas}

\item{name}{Name of the output object to fil the name slot}

\item{desc}{Description of the output object to fill the desc slot}
}
\value{
An object of class \verb{\link{FLStock}}
}
\description{
Results of a run of the Stock Synthesis sofware, SS3 (Methot & Wetzel, 2013),
can be loaded into an object of class \code{\link{FLStock}}. The code makes
use of the r4ss::SS_output function to obtain a list from Report.sso. The
following elements of that list are used to generate the necessary information
for the slots in \code{\link{FLStock}}: "catage", "natage", "ageselex",
"endgrowth", "catch_units", "nsexes", "nseasons", "nareas", "IsFishFleet",
"fleet_ID", "FleetNames", "spawnseas", "inputs" and "SS_version".
}
\references{
Methot RD Jr, Wetzel CR (2013) Stock Synthesis: A biological and statistical
framework for fish stock assessment and fishery management.
Fisheries Research 142: 86-99.
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
